Add test for the new MySQL backend
There is one known failure at the moment:

    ### Failure in: 1:MySQL backend tests
    Installed migrations
    expected: ["root"]
     but got: ["root","second"]

In this test two migrations are applied in the same transaction, the
first containing valid SQL and the second containing invalid SQL. In
PostgreSQL and SQLite, DDL is transactional and so the initial created
table is rolled back within the transaction. However, MySQL does not
support this - DDL statements aren't scoped by transactions. This means
that even though the third migration fails, the previous migrations are
still visible.