Move BackendTest back into dbmigrations package
BackendTest needs to be exported so that other packages can use it.
This means it needs to be part of the library as an exposed module.
As a consequence, the library (not the test-suite) now depend on HUnit,
which is a bit odd.

Also, this commit introduces a DatabaseType enumeration and adds an
attribute of that type to the Backend record, so that you can ask a
backend what type of database it represents. This is used in BackendTest
to avoid running tests which are expected to fail.