Decode directly to Migration type
This patch defines a `MigrationYaml` record that represents the shape of
a migration's Yaml file directly, then decodes to it. This obviates the
manual "parsing" from a decoded `Object`. It also produces better error
messages.

The tests don't pass because we have actually been using Show/Read for
the `UTCTime` fields instead of a JSON representation. I'll have to
introduce a new type to account for that without breaking compatibility.