Use custom UTCTime type to retain Show/Read format
There's no functional need for this, using the JSON representation
would've been readable enough, but there are going to be tons of
migration files in the wild with these simplified values.

We could write encoding at JSON and a decoder that accepts either, but
I'm not sure it's worth it vs just using a completely custom newtype.